add appointment

+++ src/main/java/com/clinicmanagement/clinic/repository/AppointmentRepository.java
@@ -2,7 +2,10 @@ package com.clinicmanagement.clinic.repository;
 
 import com.clinicmanagement.clinic.Entities.Appointment;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+import java.util.List;
+
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
 
 }
